Files and Folder Added
Design Folder: contain photos of the handmade design of the website. There are two photos representing two versions of the design. The one currently being used is the number 2.

Index.html: File added with the beginning of the HTML skeleton.

REAMDE: section in Brazilian Portuguese that was in English was fixed. Small content addition to the `My Process` section.