New File Addition
Three New pages created. They are broken down to make it easier to fill different parts of the form.
There is also a CSS file with a small bit of style to give the HTML pages some color when visualizing it. This file will be changed according to the project necessity.